more instructions

+++ DerpNES/Instruction.cs
@@ -1,3 +1,3 @@
 ﻿namespace DerpNES;
 
-internal record struct Instruction( uint Opcode, Func<UInt8> Operate, Func<UInt8> AddressMode, uint Cycles );
+internal record struct Instruction( uint Opcode, Func<UInt8> Operate, Func<UInt8> AddressMode, UInt8 Cycles );
